First Lecture

+++ triangle.py
@@ -4,7 +4,7 @@
 #Date:- 8/27/2020
 import math
 print("Compute the hypotenuse of right triangle")
-leg1 = int(input("Enter first length:"))
-leg2 = int(input("Enter second length:"))
+leg1 = float(input("Enter first length:"))
+leg2 = float(input("Enter second length:"))
 leg3=math.sqrt(leg1*leg1+leg2*leg2)
 print(leg3)
